Merged PR 2260: Edit team, members table
https://rctech.atlassian.net/browse/EMISV2-266

After creating the team, we are redirected to edit page within a second. Yet, the data takes more time than 1sec to be reflected in the indexes. This will need to be addressed later to see what strategy we can take.

So see the changes after edit, please refresh the page manually after some time