Merged PR 2379: Change logic for team so we use local data instead of waiting for BE
https://rctech.atlassian.net/browse/EMISV2-873

All components get team with a computed property linked to store.

**Note that we store a instance of a class directly in the store so we can leverage methods from the class itself, while keeping the logic in the class**

For instance:

  `removeTeamMember(state: IState, id: uuid) {
    state.team.removeTeamMember(id);
  },`

All actions (edit, add, remove, create) triggers API calls. We update team in store **with the payload instead of the response**. As soon as it's done, all components will be updated thanks to the computed props.

During testing, the methods of the instance are remove and only attributes are preserved. If your tests involved one of this method, we just need to force the team with a new instance.

Example:

` store.state.team.team = new Team();`