Merged PR 2859: emisv2-113 - PR for merging project split into working feature
At the beginning of this PR, the Shelter Locations feature works. The service endpoint has some variances in naming with fields such as :
- EventId
- EventName

The goal of this PR is to merge in the latest from master while splitting out the files that should be in the library.

Ticket: https://rctech.atlassian.net/browse/EMISV2-113

UPDATE: Bug that was fixed in the latest commit - https://rctech.atlassian.net/browse/EMISV2-1221