Merged PR 3027: FIX - Related events dissapearing
FIX - Related events disappearing

Issue was caused by the fact that retrieveTeamEvents was only working with Open events.

As soon as a onHold event was listed on Teams it would be undefined after editing or creating the team

To test: Take a team which is linked is an on hold event and edit anything, before it will disappear. After it won't

https://rctech.atlassian.net/browse/EMISV2-1282
https://rctech.atlassian.net/browse/EMISV2-1281
https://rctech.atlassian.net/browse/EMISV2-1277