Merged PR 3410: Canceling role assignment after previous role assignment
These changes correct the case where a user changes a UserAccount's role, sets a new role for that UserAccount and cancels the change. The key change is that when a new role is assigned, after completion of the assignment the "original users" listing is refreshed from the store (it was updated as part of the role assignment). This list is used to track the attributes for a UserAccount prior to any pending change (e.g. a new role is selected but not yet applied).

- Comment explaining issue: https://rctech.atlassian.net/browse/EMISV2-1038?focusedCommentId=23845
- IRA Ticket: https://rctech.atlassian.net/browse/EMISV2-1038

How to verify the fix:
- Assign a new role to a user in the UserAccount listing (this new role will be **Role B**) and apply it (note the role prior to the change - which we'll call **Role A**)
- Assign a new role to that same user
- Cancel the role change
- Verify that the role matches **Role B** and not **Role A**

Before the fix, the role would have matched **Role A**