Merged PR 3721: Side car user account
This is the PR for the implementation of the side-car response-handling architecture (using nested entity and metadata attributes).

By "nested", the structure of the entities has been changed from a "flattened" form where the properties for both entity and metadata are at the top-level of the interface to the following:

```
export interface IEntityCombined {
  entity: IEntity,
  metadata: IEntity,
}
```
The first extension of this is:
```
export interface IUserAccountCombined extends IEntityCombined {
  entity: IUserAccountEntity,
  metadata: IUserAccountMetadata,
}
```

The first entity type that has been brought into this approach is the UserAccount. All efforts have been made not to interfere with other entities that are derived from UserAccount (such as ITeamMember). In those cases, the entity and metadata have ben flattened to be compatible with the existing code.

Main items included in this effort:

- Create base service
- Create base module
- Create base storage combined
- User account modules for entities + metadata
- User account combined storage
- Removed user services
- Removed app-user storage and store
- Clean up user by removing filters from there
- Use new userAccount entities in teams, account, filters and case file assignment
- FIX - Account settings page

https://rctech.atlassian.net/browse/EMISV2-1456

![Screen Shot 2021-06-09 at 12.30.52 PM.png](https://dev.azure.com/CRC-Tech/5c894eca-c419-42ba-be9b-389b20bf1a87/_apis/git/repositories/b457b0f5-ff32-4bdc-b38f-50c0b2ce1a71/pullRequests/3468/attachments/Screen%20Shot%202021-06-09%20at%2012.30.52%20PM.png)
-+-+