Merged PR 7572: [EMISV2-4289] Navigation to a url that is not a GET throws an error
https://rctech.atlassian.net/browse/EMISV2-4289

The error was in the routing, because the referral details page had a similar route as the add and edit page, (but with a parameter for the referralId) when the page was reloaded with the referral/add url, it would first try to open the details page and launched a fetch call to get the data for that page, with “add” as referralId. Inversing the order of the pages in the routes list solved the issue.