Merged PR 9191: First part - Jest improvements
We are not using the package `testing-lib/js-dom` enough to keep it, so I removed it.

Also I fixed an issue where test for userStore were always triggered before each test.

This PR is the first of a series in the future that should reduce the memory usage and speed up tests (hopefully)