Merged PR 9422: [EMISV2-6463] Use event search instead of metadata for related event display
https://rctech.atlassian.net/browse/EMISV2-6463

Issue: related events in event metadata are not updated when a dependent event changes.

Instead of modifying the BE to fix this problem, the FE has been altered to get rid of the dependency on this part of event metadata. Since the related event display on the Event Details page was the only location that used this, all traces of "EventMetadata.RelatedEventsInfos" have been removed.