Merged PR 10265: [EMISV2-7018] - Better handling for errors while loading the app
Updated LoginError component so we can pass a titleKey and messageKey.

When signing in with an account that is in AAD but does not have access to our tenant, I encounter two scenarios:
- Issues while fetching users
- Issues while fetching tenant settings

Took care of those cases by throwing errors, catching them, and setting custom messages so we avoid INC tickets.

**One issue I could not solve**: once on the login error page, if the user clicks on refresh, we have an error with Pinia saying it is not loaded yet.