Merged PR 11198: [EMISV2-7577] notification links
https://rctech.atlassian.net/browse/EMISV2-7577

Link notification subject text to navigate to target entities when:
- the notification type supports it (e.g. Tasks)
- the target entity can be loaded

Note: loading the target entity from the FE may seem suboptimal, but it will be required for display of current entity state (e.g. Urgent, Due Today, Past Due, etc.). Doing this in the BE was considered, but doing so would have required reimplementation of authorization logic for any targeted entity. Our pattern is to share data between domains by making repositories accessible everywhere, but this shares no authorization logic. An alternative might also have been to make API calls from the BE, but this would require new infrastructure that is already in the FE.