Merged PR 11308: [EMISV2-7934] [EMISV2-7937] Get entities from API by list of ids
https://rctech.atlassian.net/browse/EMISV2-7934

Note: branch name is subtask due to TF problems with 7934 FB.

Entity controllers in the API have modified resource GETs that take an optional list of ids. Support for this has been added to the DomainBaseService as well as the base store components. Note that controllers for "child" entities do not have this modification by default as it may not make sense. Specific implementation has been added for Tasks as it falls into the "child" entity category.

The notification center has also been reworked to use this method of fetching linked entities. When notifications are fetched, any supported target entities are subsequently fetched. The NotificationCard component then simply binds to target entities, if they are available.